third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-with-host-pseudo-class-expected.txt

CONSOLE WARNING: line 34: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
content

Test for crbug.com/274063, cannot style ::content with a rule that includes :host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(document.querySelector("#host > .content")).color is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
